Refactor data setup withSeller/Product (bug 938658)
