+++ test/suite/payments.test.js
@@ -1,19 +1,18 @@
 var url = require('url');
 
 var under = require('underscore');
-var uuid = require('node-uuid');
 var supertest = require('super-request');
 
 var test = require('../');
 var Client = require('../client').AnonymousClient;
-var products = require('../../lib/products');
-var sellers = require('../../lib/sellers');
+var helpers = require('../helpers');
 var trans = require('../../lib/trans');
 
 var client = new Client('/');
 
 
 var transData = {
+  /*jshint camelcase: false */
   product_id: undefined,
   region: 123,
   carrier: 'USA_TMOBILE',
@@ -32,34 +31,10 @@ var newTransData = under.extend({}, transData, {
 });
 
 
-function withSeller(opt, cb) {
-  opt = opt || {};
-  var props = under.extend({uuid: uuid.v4(), status: 'ACTIVE'}, opt);
-  sellers.models.create(props, function(err, seller) {
-    if (err) {
-      throw err;
-    }
-    cb(seller);
-  });
-}
-
-
-function withProduct(opt, cb) {
-  opt = opt || {};
-  var props = under.extend({external_id: uuid.v4(), status: 'ACTIVE'}, opt);
-  products.models.create(props, function(err, product) {
-    if (err) {
-      throw err;
-    }
-    cb(product);
-  });
-}
-
-
 exports.setUp = function(done) {
   trans.models.deleteMany({}, function() {
-    withSeller({}, function(seller) {
-      withProduct({seller_id: seller._id}, function(product) {
+    helpers.withSeller(undefined, {}, function(seller) {
+      helpers.withProduct(undefined, {seller_id: seller._id}, function(product) {
         transData.product_id = product._id;
         trans.models.create(transData, function(err, createdTrans) {
           if (err) {
@@ -78,7 +53,7 @@ exports.testStartTransThenProcess = function(t) {
   supertest(test.app)
     .get('/?tx=' + transData.token)
     .expect(200)
-    .end(function(err, res) {
+    .end(function(err) {
       t.ifError(err);
     })
     .post('/payment/process')
@@ -99,7 +74,7 @@ exports.testStartTransThenFail = function(t) {
   supertest(test.app)
     .get('/?tx=' + transData.token)
     .expect(200)
-    .end(function(err, res) {
+    .end(function(err) {
       t.ifError(err);
     })
     .post('/payment/process')
@@ -126,7 +101,7 @@ exports.testNoActiveTrans = function(t) {
   supertest(test.app)
     .post('/payment/process')
     .expect(409)
-    .end(function(err, res) {
+    .end(function(err) {
       t.ifError(err);
       t.done();
     });
@@ -135,15 +110,8 @@ exports.testNoActiveTrans = function(t) {
 
 function createTrans(done, params) {
   params = params || {};
-  var t = {
-    ifError: function(err) {
-      if (err) {
-        throw err;
-      }
-    }
-  }
-  withSeller({}, function(seller) {
-    withProduct({seller_id: seller._id, active: false},
+  helpers.withSeller(undefined, {}, function(seller) {
+    helpers.withProduct(undefined, {seller_id: seller._id, active: false},
       function(product) {
         var data = under.extend(newTransData, {
           product_id: product._id,
@@ -155,14 +123,14 @@ function createTrans(done, params) {
           console.log('created trans', trans);
           done(trans);
         });
-      })
+      });
   });
 }
 
 
 exports.testNoToken = function(t) {
   client.get().expect(409)  // missing tx=
-    .end(function(err, res) {
+    .end(function(err) {
       t.ifError(err);
       t.done();
     });
@@ -171,7 +139,7 @@ exports.testNoToken = function(t) {
 
 exports.testInvalidToken = function(t) {
   client.get({tx: 'nope'}).expect(404)
-    .end(function(err, res) {
+    .end(function(err) {
       t.ifError(err);
       t.done();
     });
@@ -181,7 +149,7 @@ exports.testInvalidToken = function(t) {
 exports.testGoodToken = function(t) {
   createTrans(function(trans) {
     client.get({tx: trans.token}).expect(200)
-      .end(function(err, res) {
+      .end(function(err) {
         t.ifError(err);
         t.done();
       });
@@ -192,7 +160,7 @@ exports.testGoodToken = function(t) {
 exports.testEndedTrans = function(t) {
   createTrans(function(trans) {
     client.get({tx: trans.token}).expect(400)
-      .end(function(err, res) {
+      .end(function(err) {
         t.ifError(err);
         t.done();
       });
