+++ test/suite/state.test.js
@@ -1,5 +1,4 @@
 var Client = require('../client').Client;
-var state = require('../../lib/state');
 
 var client = new Client('/status');
 
@@ -9,6 +8,6 @@ exports.testOK = function(t) {
     .expect(200)
     .end(function(err, res) {
       t.equal(res.body.result, 'OK');
-      t.done()
+      t.done();
     });
-}
+};
